Refactor devtools setup and remove unused code
